migrate(user-id): from 32-bit integer to 64-bit integer

+++ include/tgbot/types/ResponseParameters.h
@@ -22,7 +22,7 @@ public:
      * difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a
      * signed 64 bit integer or double-precision float type are safe for storing this identifier.
      */
-    std::int32_t migrateToChatId;
+    std::int64_t migrateToChatId;
 
     /**
      * @brief Optional. In case of exceeding flood control, the number of seconds left to wait before the request can be repeated
